Remove dependency on goog.debug.Logger
The GClosure logging library is slightly restrictive in that it can only
log a single (string) argument per message. This is fine when we only
want to print a simple string representation, but developers may want to
use utilities like [cljs-devtools][cljs-devtools] to inspect logged
objects.

To support this usecase, we need to be able pass objects directly to
`console`, unformatted, and so this commit removes the dependency on
goog.debug.Logger.

To support this, I added a small set of configuration options and
implemented a basic logger hierarchy. The configuration defaults mostly
match the original API.

One notable backward-incompatible difference is that the timestamp is no
longer emitted in log messages, which is something we could add back if
needed.

 [cljs-devtools]: https://github.com/binaryage/cljs-devtools